# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sample Swagger",
        "description": "Example",
        "version": "1.0.0"
    },
    "paths": {
        "/customers/create-new": {
            "post": {
                "tags": [
                    "Example"
                ],
                "summary": "Create",
                "description": "New User",
                "operationId": "create",
                "requestBody": {
                    "description": "blabla",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Dela Cruz"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "09178882222"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Created Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to create User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/update": {
            "put": {
                "tags": [
                    "Example"
                ],
                "summary": "Update",
                "description": "Update User",
                "operationId": "Update",
                "requestBody": {
                    "description": "blabla",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Juan"
                                    },
                                    "last": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "Dela Cruz"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "format": "string",
                                        "example": "09178882222"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Updated Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to update User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/delete": {
            "delete": {
                "tags": [
                    "Example"
                ],
                "summary": "Delete",
                "description": "Delete User",
                "operationId": "delete",
                "requestBody": {
                    "description": "blabla",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Deleted Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to delete User",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/show": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Show",
                "description": "Show User",
                "operationId": "show",
                "requestBody": {
                    "description": "blabla",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "format": "int64",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "first": {
                                            "type": "string"
                                        },
                                        "last": {
                                            "type": "string"
                                        },
                                        "phone": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "User not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/show-all": {
            "get": {
                "tags": [
                    "Example"
                ],
                "summary": "Show All",
                "description": "Show All Users",
                "operationId": "showAll",
                "responses": {
                    "200": {
                        "description": "Users Retrieved Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "first": {
                                                "type": "string"
                                            },
                                            "last": {
                                                "type": "string"
                                            },
                                            "phone": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Users not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}